move expunge out
